feat: setup module
