+++ auth/src/main/java/com/lxp/auth/AuthApplication.java
@@ -1,0 +1,13 @@
+package com.lxp.auth;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AuthApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(AuthApplication.class, args);
+    }
+
+}
