+++ auth/src/test/java/com/lxp/auth/AuthApplicationTests.java
@@ -1,0 +1,13 @@
+package com.lxp.auth;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AuthApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
