--- auth/src/test/java/com/ogirafferes/auth/AuthApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ogirafferes.auth;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AuthApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
