+++ content/src/main/java/com/lxp/content/ContentApplication.java
@@ -1,0 +1,13 @@
+package com.lxp.content;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ContentApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ContentApplication.class, args);
+    }
+
+}
