--- content/src/main/java/com/ogirafferes/content/ContentApplication.java
@@ -1,13 +1,0 @@
-package com.ogirafferes.content;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ContentApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ContentApplication.class, args);
-    }
-
-}
