+++ content/src/test/java/com/lxp/content/ContentApplicationTests.java
@@ -1,0 +1,13 @@
+package com.lxp.content;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ContentApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
