--- content/src/test/java/com/ogirafferes/content/ContentApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ogirafferes.content;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ContentApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
