+++ enrollment/src/main/java/com/lxp/enrollment/EnrollmentApplication.java
@@ -1,0 +1,13 @@
+package com.lxp.enrollment;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EnrollmentApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(EnrollmentApplication.class, args);
+    }
+
+}
