--- enrollment/src/main/java/com/ogirafferes/enrollment/EnrollmentApplication.java
@@ -1,13 +1,0 @@
-package com.ogirafferes.enrollment;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EnrollmentApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(EnrollmentApplication.class, args);
-    }
-
-}
