+++ enrollment/src/test/java/com/lxp/enrollment/EnrollmentApplicationTests.java
@@ -1,0 +1,13 @@
+package com.lxp.enrollment;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EnrollmentApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
