--- enrollment/src/test/java/com/ogirafferes/enrollment/EnrollmentApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ogirafferes.enrollment;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EnrollmentApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
