+++ recommand/src/main/java/com/lxp/recommand/RecommandApplication.java
@@ -1,0 +1,13 @@
+package com.lxp.recommand;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RecommandApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(RecommandApplication.class, args);
+    }
+
+}
