--- recommand/src/main/java/com/ogirafferes/recommand/RecommandApplication.java
@@ -1,13 +1,0 @@
-package com.ogirafferes.recommand;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RecommandApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(RecommandApplication.class, args);
-    }
-
-}
