+++ recommand/src/test/java/com/lxp/recommand/RecommandApplicationTests.java
@@ -1,0 +1,13 @@
+package com.lxp.recommand;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class RecommandApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
