--- recommand/src/test/java/com/ogirafferes/recommand/RecommandApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ogirafferes.recommand;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RecommandApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
