+++ skill/src/main/java/com/lxp/skill/SkillApplication.java
@@ -1,0 +1,13 @@
+package com.lxp.skill;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SkillApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SkillApplication.class, args);
+    }
+
+}
