--- skill/src/main/java/com/ogirafferes/skill/SkillApplication.java
@@ -1,13 +1,0 @@
-package com.ogirafferes.skill;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SkillApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SkillApplication.class, args);
-    }
-
-}
