+++ skill/src/main/resources/application.yaml
