+++ skill/src/test/java/com/lxp/skill/SkillApplicationTests.java
@@ -1,0 +1,13 @@
+package com.lxp.skill;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SkillApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
