--- skill/src/test/java/com/ogirafferes/skill/SkillApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ogirafferes.skill;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SkillApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
