--- src/main/java/com/ogirafferes/lxpproject3rd/LxpProject3rdApplication.java
@@ -1,13 +1,0 @@
-package com.ogirafferes.lxpproject3rd;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LxpProject3rdApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(LxpProject3rdApplication.class, args);
-    }
-
-}
