--- src/test/java/com/ogirafferes/lxpproject3rd/LxpProject3rdApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ogirafferes.lxpproject3rd;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LxpProject3rdApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
