+++ tag/src/main/java/com/lxp/tag/TagApplication.java
@@ -1,0 +1,13 @@
+package com.lxp.tag;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TagApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TagApplication.class, args);
+    }
+
+}
