--- tag/src/main/java/com/ogirafferes/tag/TagApplication.java
@@ -1,13 +1,0 @@
-package com.ogirafferes.tag;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TagApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TagApplication.class, args);
-    }
-
-}
