+++ tag/src/test/java/com/lxp/tag/TagApplicationTests.java
@@ -1,0 +1,13 @@
+package com.lxp.tag;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TagApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
