--- tag/src/test/java/com/ogirafferes/tag/TagApplicationTests.java
@@ -1,13 +1,0 @@
-package com.ogirafferes.tag;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TagApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
