+++ user/src/test/java/com/lxp/user/UserApplicationTests.java
@@ -1,0 +1,13 @@
+package com.lxp.user;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
